Add files via upload
it can useful for getting first explanation of the element which you are researching from wikipedi but it's not fine enough for now . and for turkish language but i explained in english too in the code page. I m sure maybe there is the preject more useful than mine. but its my first project , my little baby... it is so special and will stay that for me